extract getStartUrl hook from site migration
